requirements.txt
# ── Core ──────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── System utilities ──────────────────────────────────
psutil>=5.9.0
Pillow>=10.3.0
# keras-ocr/imgaug is not fully NumPy 2.x compatible.
numpy>=1.26.0,<2.0.0
pdf2image

# ── Tier 2: Traditional OCR engines ───────────────────
# Install paddlepaddle/paddlepaddle-gpu separately based on your CUDA/OS from:
# https://www.paddlepaddle.org.cn/install/quick
paddleocr>=3.0.0
easyocr>=1.7.1
pytesseract>=0.3.13         # Requires: sudo apt install tesseract-ocr
keras-ocr>=0.9.3
calamari-ocr>=2.1.0
kraken>=4.3.0
mmocr>=1.0.1

# ── Tier 1: VLM OCR (uncomment when GPU available) ────
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.44.0
accelerate>=0.31.0
qwen-vl-utils>=0.0.5      # for Qwen2.5-VL
bitsandbytes      # optional 4-bit Qwen loading

# ── Dev / Testing ─────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0               # async test client for FastAPI
tiktoken
verovio
# MMOCR dependency set (mmocr 1.x requires mmcv 2.x)
mmcv>=2.0.0rc4,<2.2.0
mmdet>=3.0.0,<3.4.0
mmengine>=0.7.0,<1.0.0